graphicstextitem: Move 'NameTextItem' to separate module
